*Clean up some misc things. *Moved server request to a new class. *Fixed gradle pathing.
