+++ gradle.properties
@@ -1,6 +1,5 @@
 # Done to increase the memory available to gradle.
 org.gradle.jvmargs=-Xmx1G
-org.gradle.java.home=C:/Users/onebe/.jdks/temurin-17.0.4.1
 
 # Fabric Properties
 	# check these on https://fabricmc.net/develop
