+++ src/main/java/net/onebeastofchris/geyserplayerheads/GeyserPlayerHeads.java
@@ -1,31 +1,26 @@
 package net.onebeastofchris.geyserplayerheads;
 
-import com.google.gson.*;
 import net.fabricmc.api.ModInitializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.io.IOException;
-import java.net.URI;
-import java.net.http.HttpRequest;
-import java.net.http.HttpResponse;
-import java.net.http.HttpResponse.BodyHandlers;
-import static java.net.http.HttpClient.newHttpClient;
 
 public class GeyserPlayerHeads implements ModInitializer {
 	// This logger is used to write text to the console and the log file.
 	// It is considered best practice to use your mod id as the logger's name.
 	// That way, it's clear which mod wrote info, warnings, and errors.
-	public static final Logger LOGGER = LoggerFactory.getLogger("gph");
+	private static Logger logger;
 
 	@Override
 	public void onInitialize() {
 		// This code runs as soon as Minecraft is in a mod-load-ready state.
 		// However, some things (like resources) may still be uninitialized.
 		// Proceed with mild caution.
-
-		LOGGER.info("GeyserPlayerHeads starting now");
+		logger = LoggerFactory.getLogger("gph");
+		logger.info("GeyserPlayerHeads starting now");
 	}
 
+	public static Logger getLogger() {
+		return logger;
+	}
 }
